requirements.txt
mock
pytest-django
statsd==3.2.1
flake8
pep8
pyflakes
mccabe
pep8-naming
flake8-print
flake8-debugger
